clang-tidy sweep: floorf/fabsf, exact string match, branch consolidation
- Use floorf/fabsf on float positions instead of double floor/fabs.
- Replace strncmp(..., N) shadeType matching with exact strcmp.
- Collapse redundant if/else chains in GPIO direction and pin checks.
- Simplify RX bit_length sync-pulse classification.
- Self-contain SomfyTargetSequencer.h with stdint and SomfyFrame includes.
- Mark unused-in-non-DEBUG `timing` local as [[maybe_unused]].

+++ main/somfy/shade/SomfyCommandProcessor.cpp
@@ -24,7 +24,7 @@ void SomfyCommandProcessor::processWaitingFrame() {
           shade->lastFrame.processed = true;
           // Simply move the shade up by 1%.
           if(shade->currentPos > 0) {
-            shade->p_target(floor(shade->currentPos) - 1);
+            shade->p_target(floorf(shade->currentPos) - 1);
             shade->setMovement(-1);
             shade->emitCommand(cmd, "remote", shade->lastFrame.remoteAddress);
           }
@@ -33,7 +33,7 @@ void SomfyCommandProcessor::processWaitingFrame() {
           shade->lastFrame.processed = true;
           // Simply move the shade down by 1%.
           if(shade->currentPos < 100) {
-            shade->p_target(floor(shade->currentPos) + 1);
+            shade->p_target(floorf(shade->currentPos) + 1);
             shade->setMovement(1);
             shade->emitCommand(cmd, "remote", shade->lastFrame.remoteAddress);
           }
@@ -86,7 +86,7 @@ void SomfyCommandProcessor::processWaitingFrame() {
         break;
       case somfy_commands::My:
         if(shade->lastFrame.repeats >= SETMY_REPEATS && shade->isIdle()) {
-          if(floor(shade->getMyPos()) == floor(shade->currentPos)) {
+          if(floorf(shade->getMyPos()) == floorf(shade->currentPos)) {
             // We are clearing it.
             shade->p_myPos(-1);
             shade->p_myTiltPos(-1);
